refactor: move validation onto utils folder
