+++ client/src/features/Home/Searchbar.tsx
@@ -5,14 +5,9 @@ import ToggleButton from '@mui/material/ToggleButton';
 import Box from '@mui/material/Box';
 import { Controller, useForm } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
-import * as yup from 'yup';
 import { useAppDispatch } from 'app/hooks';
 import { searchActions, YoutubeSearchForm } from 'slices/search/slice';
-
-export const youtubeSearchSchema = yup.object({
-  searchTerm: yup.string().required(),
-  type: yup.string().required(),
-});
+import { youtubeSearchSchema } from 'utils/validation.util';
 
 export function Searchbar() {
   const dispatch = useAppDispatch();
@@ -22,7 +17,9 @@ export function Searchbar() {
     resolver: yupResolver(youtubeSearchSchema),
   });
 
-  const onSubmit = form.handleSubmit(formData => dispatch(searchActions.searchYoutube(formData)));
+  const onSubmit = form.handleSubmit(formData =>
+    dispatch(searchActions.searchYoutube(formData)),
+  );
 
   return (
     <Box component="form" display="flex" justifyContent="center" onSubmit={onSubmit}>
@@ -40,7 +37,14 @@ export function Searchbar() {
         control={form.control}
         name="searchTerm"
         render={({ field, fieldState: { error } }) => (
-          <TextField label="Course Subject" error={!!error} helperText={error?.message} fullWidth {...field} sx={{ width: '40%', mr: 1 }} />
+          <TextField
+            label="Course Subject"
+            error={!!error}
+            helperText={error?.message}
+            fullWidth
+            {...field}
+            sx={{ width: '40%', mr: 1 }}
+          />
         )}
       />
       <Button size="large" color="primary" variant="contained" type="submit">
